Move SquatchAndroid to interface + cleanup
